# web/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Socle Symfony2 PFE API",
        "description": "Documented API for Socle symfony2",
        "version": "1.0.0"
    },
    "host": "socle-symfony2-pfe.local",
    "basePath": "/app_dev.php",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "getUsers"
                ],
                "summary": "users list",
                "operationId": "showUsersList",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Droit non attribue pour cet utilisateur"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "getUserById"
                ],
                "summary": "Info for a specific user",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the user to retrieve",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "api/users/{username}": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Delete a specific user",
                "description": "Delete a specific user.",
                "operationId": "deleteUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The username that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "properties": {
                "id": {
                    "description": "$id",
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                }
            },
            "type": "object"
        }
    }
}